simplify examples
