+++ Examples/Examples/01-DefaultSankeyView.swift
@@ -1,13 +1,14 @@
 import Sankey
+import SwiftUI
 
-extension SankeyData {
-    static let preview = SankeyData(
+struct DefaultSankeyView: View {
+    let data = SankeyData(
         nodes: [
             SankeyNode("A"),
             SankeyNode("B"),
             SankeyNode("X"),
             SankeyNode("Y"),
-            SankeyNode("Z"),
+            SankeyNode("Z")
         ],
         links: [
             SankeyLink(5, from: "A", to: "X"),
@@ -15,7 +16,17 @@ extension SankeyData {
             SankeyLink(6, from: "A", to: "Z"),
             SankeyLink(2, from: "B", to: "X"),
             SankeyLink(9, from: "B", to: "Y"),
-            SankeyLink(4, from: "B", to: "Z"),
+            SankeyLink(4, from: "B", to: "Z")
         ]
     )
+    
+    var body: some View {
+        SankeyDiagram(data)
+            .padding(10)
+            .frame(height: 250)
+    }
+}
+
+#Preview {
+    DefaultSankeyView()
 }
