--- Examples/Examples/Views/DefaultSankeyView.swift
@@ -1,16 +1,0 @@
-import Sankey
-import SwiftUI
-
-struct DefaultSankeyView: View {
-    let data: SankeyData = .preview
-    
-    var body: some View {
-        SankeyDiagram(data)
-            .padding(10)
-            .frame(height: 250)
-    }
-}
-
-#Preview {
-    DefaultSankeyView()
-}
